feat(ansible): add Homebrew taps and options support for MacOS
Added support for Homebrew taps and install options in the package generator.
The `PackageData` struct now includes new fields for `brewTap` and
`brewOptions`. Updated the templates and playbooks to utilize these fields.

Examples of usage include specifying a tap (e.g., `d12frosted/emacs-plus`) and
passing installation options such as `with-native-comp` and `with-modern-icon`
for packages installed on MacOS via Homebrew.