fix(ansible): Handle command names starting with numbers
Ansible variables cannot start with a number. This commit updates the
`generate_packages.go` script to prefix commands that start with a
number with `cmd_`.

This commit also introduces a new `Makefile` in the `devtools/setup-dev/ansible`
directory to make it easier to regenerate the Ansible playbooks. The root
`Makefile` has been updated to call this new `Makefile`.